So I had finish my project as you can see my code is in mytcp, but I wasn;t able to run it. So I asked claude for help and asked to fix my code but in another directory (tcp_working) in this direc the code is worknig, so I will use it for future debug of my repo and try it to fix myself

+++ projects/mytcp/shared/pseIPHeader.h
@@ -9,7 +9,7 @@ typedef struct PseudoIpHeader{
     uint32_t ipv4_dest;
     uint8_t  zero_byte;
     uint8_t  protocol_num; 
-    uint16_t tcp_segment_lenght; /* BUG FIX: was uint8_t, RFC 793 pseudo-header needs 16 bits */
+    uint8_t  tcp_segment_lenght;
 } PseudoIpHeader;
 
 PseudoIpHeader *psdIpHeader_intit(uint32_t ipv4_scr,
